Summary: Add in loading progress bar
Test Plan:

Reviewers:

Subscribers: samz, kelseylam, baerica, amillman

Projects: #Sage-Mobile
