+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/checkIn/fragments/CreateCheckInFragment.java
@@ -29,6 +29,7 @@ import java.util.Calendar;
 
 import blueprint.com.sage.R;
 import blueprint.com.sage.events.checkIns.CheckInEvent;
+import blueprint.com.sage.models.APIError;
 import blueprint.com.sage.models.CheckIn;
 import blueprint.com.sage.models.School;
 import blueprint.com.sage.models.User;
@@ -66,6 +67,8 @@ public class CreateCheckInFragment extends Fragment implements FormValidation, D
 
     private BaseInterface mBaseInterface;
     private ToolbarInterface mToolbarInterface;
+
+    private MenuItem mItem;
 
     public static CreateCheckInFragment newInstance() { return new CreateCheckInFragment(); }
 
@@ -107,6 +110,7 @@ public class CreateCheckInFragment extends Fragment implements FormValidation, D
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
+        mItem = item;
         switch (item.getItemId()) {
             case R.id.menu_save:
                 validateAndSubmitRequest();
@@ -223,6 +227,7 @@ public class CreateCheckInFragment extends Fragment implements FormValidation, D
             Snackbar.make(mLayout, R.string.check_in_request_blank, Snackbar.LENGTH_SHORT).show();
         }
 
+        mItem.setActionView(R.layout.actionbar_indeterminate_progress);
         CheckIn checkIn = new CheckIn(startDate.toDate(), endDate.toDate(), user, school, comment);
         Requests.CheckIns.with(getActivity()).makeCreateRequest(checkIn);
     }
@@ -240,6 +245,7 @@ public class CreateCheckInFragment extends Fragment implements FormValidation, D
     }
 
     public void onEvent(CheckInEvent event) { resetCheckIn(); }
+    public void onEvent(APIError event) { mItem.setActionView(null); }
 
     /**
      * Pickers for both date and time
