+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/users/profile/fragments/EditUserFragment.java
@@ -3,6 +3,7 @@ package blueprint.com.sage.users.profile.fragments;
 import android.util.Log;
 import android.view.View;
 
+import blueprint.com.sage.R;
 import blueprint.com.sage.events.users.EditUserEvent;
 import blueprint.com.sage.models.School;
 import blueprint.com.sage.models.User;
@@ -53,6 +54,7 @@ public class EditUserFragment extends UserFormAbstractFragment {
         mUser.setSchoolId(schoolId);
         mUser.setProfile(mProfileBitmap);
 
+        mItem.setActionView(R.layout.actionbar_indeterminate_progress);
         Requests.Users.with(getActivity()).makeStickyEditRequest(mUser);
     }
